Fix destination option
If the destination folder is set, it will be repeated on save. (e.g. if the destination folder is set to `example/`, the file will be saved to `example/file.ext` but will be saved by Ghost as `example/example/file.ext`.) This change fixes that by referencing the file name only.